Move attachment handler to separate file
